[fix] add buff

+++ examples/uart_tanslation_i2c/translation/buff.h
@@ -8,7 +8,7 @@ void buf_push(uint8_t data);    //入队列
 uint8_t buf_pop();              //出队列
 
 
-void buf_clean(uint8_t num);
+void buf_clean(uint32_t num);
 void buf_switch(int flag);
 
 #endif
